Mini Update Nesting Checks
Mini update containing nesting for all of our checks. the storage tests could use further separation with nesting but have generally left this as not currently looking through them in detail.
