+++ src/test/java/com/sparta/group2/controller/FactoryControllerChecks.java
@@ -4,6 +4,9 @@ import org.junit.jupiter.api.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class FactoryControllerChecks {
+    @Nested
+    @DisplayName("Factory controller validation checks:")
+    public class FactoryControllerValidationChecks {
 
-
+    }
 }
