+++ src/test/java/com/sparta/group2/controller/SLQControllerChecks.java
@@ -4,6 +4,9 @@ import org.junit.jupiter.api.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class SLQControllerChecks {
+    @Nested
+    @DisplayName("SQL controller checks:")
+    public class SQLControllerChecks {
 
-
+    }
 }
